IDEMPIERE-1146 + IDEMPIERE-3234 (#76)
Avoid logging the update to ad_sequence caused by the previous fixes for IDEMPIERE-1146 + IDEMPIERE-3234

+++ org.adempiere.base/src/org/compiere/dbPort/Convert.java
@@ -549,6 +549,9 @@ public abstract class Convert
 			return true;
 		if (uppStmt.startsWith("UPDATE R_REQUESTPROCESSOR SET DATELASTRUN"))
 			return true;
+		// don't log sequence updates
+		if (uppStmt.startsWith("UPDATE AD_SEQUENCE SET CURRENTNEXT"))
+			return true;
 		// Don't log DELETE FROM Some_Table WHERE AD_Table_ID=? AND Record_ID=?
 		if (uppStmt.startsWith("DELETE FROM ") && uppStmt.endsWith(" WHERE AD_TABLE_ID=? AND RECORD_ID=?"))
 			return true;
